backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
email-validator
python-jose
passlib[bcrypt]
bcrypt<4.0
python-multipart
httpx
pytest
